Fix the build on Windows and the less common Unixes
syscall.Statfs does not exist on Windows, so the first push broke the
cross-compile job. Disk measurement is now per-platform: statfs on
linux/darwin/freebsd, GetDiskFreeSpaceEx on Windows, and a fallback
elsewhere that reports the volume as not visible rather than inventing a
number. OpenBSD spells the statfs fields differently and takes the
fallback.

Field types differ across Unixes — Bavail is signed on FreeBSD — so the
statfs path converts rather than assumes.

The cross-compile matrix now covers windows/arm64, freebsd and openbsd
too, since the gap is what let this through.

+++ internal/ui/diskspace_statfs.go
@@ -1,3 +1,5 @@
+//go:build linux || darwin || freebsd
+
 package ui
 
 import "syscall"
@@ -15,9 +17,12 @@ func volumeUsage(path string) (free, total uint64, ok bool) {
 	if err := syscall.Statfs(path, &st); err != nil {
 		return 0, 0, false
 	}
+	// Field types differ across Unixes (Bavail is signed on FreeBSD), so
+	// convert rather than assume.
 	bs := uint64(st.Bsize)
-	total = st.Blocks * bs
-	free = st.Bavail * bs
+	total = uint64(st.Blocks) * bs
+	// Bavail, not Bfree: the reserved blocks are not ours to fill.
+	free = uint64(st.Bavail) * bs
 	if total == 0 {
 		return 0, 0, false
 	}
